improve: @roots/bud-server dependency tree
