+++ packages/@roots/bud-server/package.json
@@ -63,16 +63,18 @@
     "pkg:format": "yarn g:pkg:format"
   },
   "devDependencies": {
-    "@roots/bud-framework": "workspace:packages/@roots/bud-framework",
+    "@types/express": "^4",
     "@types/node": "16.11.9",
     "@types/webpack-dev-middleware": "5.0.2",
     "@types/webpack-hot-middleware": "2.25.5",
     "webpack": "5.64.2"
   },
   "dependencies": {
+    "@roots/bud-framework": "workspace:packages/@roots/bud-framework",
     "@roots/bud-support": "workspace:packages/@roots/bud-support",
     "@roots/container": "workspace:packages/@roots/container",
     "chokidar": "^3.5.2",
+    "express": "^4.17.1",
     "http-proxy-middleware": "2.0.1",
     "lodash.isequal": "^4.5.0",
     "tslib": "^2.3.1",
