--- packages/@roots/bud-server/src/Server/server.dependencies.ts
@@ -1,11 +1,0 @@
-import chokidar from 'chokidar'
-export {chokidar}
-
-import {globby} from '@roots/bud-support'
-export {globby}
-
-import {lodash} from '@roots/bud-support'
-export const {isNull, isUndefined} = lodash
-
-import {bind, fs} from '@roots/bud-support'
-export {bind, fs}
