--- packages/@roots/bud-server/src/client/ErrorOverlay/template.ts
@@ -1,5 +1,0 @@
-export const template = content => `
-<bud-overlay>
-  ${content}
-</bud-overlay>
-`
