+++ packages/@roots/bud-server/src/client/overlay/template.web-component.ts
@@ -1,0 +1,2 @@
+export const template = content =>
+  `<bud-overlay>${content}</bud-overlay>`
