+++ packages/@roots/bud-server/src/index.ts
@@ -14,4 +14,4 @@
  * @core @packageDocumentation
  */
 
-export {Server} from './Server'
+export {Server} from './server'
