+++ packages/@roots/bud-support/package.json
@@ -69,7 +69,6 @@
     "@endemolshinegroup/cosmiconfig-typescript-loader": "3.0.2",
     "@types/copy-webpack-plugin": "8.0.1",
     "@types/express": "4.17.13",
-    "@types/jest": "27.0.3",
     "@types/node": "16.11.9",
     "@types/prettier": "2.4.2",
     "@types/safe-json-stringify": "1.1.2",
@@ -81,12 +80,10 @@
     "dotenv": "10.0.0",
     "dotenv-expand": "5.1.0",
     "execa": "5.1.1",
-    "express": "4.17.1",
     "fs-extra": "10.0.0",
     "globby": "11.0.4",
     "helpful-decorators": "2.1.0",
     "human-readable": "0.2.1",
-    "jest": "27.3.1",
     "json5": "2.2.0",
     "lodash": "4.17.21",
     "nanoid": "3.1.30",
@@ -99,12 +96,9 @@
     "signale": "1.4.0",
     "strip-ansi": "7.0.1",
     "toml": "3.0.0",
-    "ts-node": "10.4.0",
     "tslib": "2.3.1",
     "typescript": "4.4.4",
-    "vfile-reporter": "7.0.2",
     "webpack": "5.64.2",
-    "webpack-config-dump-plugin": "3.0.1",
     "yamljs": "0.3.0"
   },
   "stableVersion": "4.6.0"
