--- packages/@roots/bud-support/src/external/express.ts
@@ -1,2 +1,0 @@
-import Express from 'express'
-export {Express}
