+++ packages/@roots/bud-support/src/index.ts
@@ -34,7 +34,6 @@ export * from './external/cosmiconfig'
 export * from './external/cli-highlight'
 export * from './external/dotenv'
 export * from './external/dotenv-expand'
-export * from './external/express'
 export * from './external/pretty-format'
 export * from './external/fs-extra'
 export * from './external/globby'
