--- packages/@roots/bud/src/services/Server/server.dependencies.ts
@@ -1,2 +1,0 @@
-export {Express as Application} from '@roots/bud-support'
-export {Server as Base} from '@roots/bud-server'
