--- packages/@roots/bud/src/services/Server/server.interface.ts
@@ -1,2 +1,0 @@
-import {Server as Base} from '@roots/bud-server'
-export type {Base}
