+++ packages/@roots/bud/tsconfig-esm.json
@@ -6,40 +6,40 @@
   },
   "references": [
     {
-      "path": "./../bud-framework/tsconfig-esm.json"
+      "path": "../bud-framework/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-api/tsconfig-esm.json"
+      "path": "../bud-api/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-build/tsconfig-esm.json"
+      "path": "../bud-build/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-cache/tsconfig-esm.json"
+      "path": "../bud-cache/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-compiler/tsconfig-esm.json"
+      "path": "../bud-compiler/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-dashboard/tsconfig-esm.json"
+      "path": "../bud-dashboard/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-extensions/tsconfig-esm.json"
+      "path": "../bud-extensions/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-hooks/tsconfig-esm.json"
+      "path": "../bud-hooks/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-server/tsconfig-esm.json"
+      "path": "../bud-server/tsconfig-esm.json"
     },
     {
-      "path": "./../bud-support/tsconfig-esm.json"
+      "path": "../bud-support/tsconfig-esm.json"
     },
     {
-      "path": "./../container/tsconfig-esm.json"
+      "path": "../container/tsconfig-esm.json"
     },
     {
-      "path": "./../dependencies/tsconfig-esm.json"
+      "path": "../dependencies/tsconfig-esm.json"
     }
   ]
 }
